Enable watch and rebuild for easier dev

+++ tasks/polyfillservice.js
@@ -4,12 +4,17 @@ module.exports = function(grunt) {
 	var spawn = require('child_process').spawn;
 	var path = require('path');
 	var serviceModule = path.join(__dirname, '/../service/index.js');
+	var serviceProcess = null;
 
 	grunt.registerTask('polyfillservice', "Start the polyfill service", function() {
 
 		var done = this.async();
 
-		var serviceProcess = spawn(process.argv[0], [serviceModule], { silent: true });
+		if (serviceProcess) {
+			serviceProcess.kill();
+		}
+
+		serviceProcess = spawn(process.argv[0], [serviceModule], { silent: true, env: process.env, cwd: path.join(__dirname, '/..')});
 
 		serviceProcess.stdout.on('data', function(d) {
 			var data = d.toString();
@@ -19,9 +24,14 @@ module.exports = function(grunt) {
 			}
 		});
 
-		process.on('exit', function() {
-			serviceProcess.kill();
+		serviceProcess.on('close', function (code, signal) {
+			console.log('child process terminated due to receipt of signal '+signal);
 		});
+	});
 
+	process.on('exit', function() {
+		if (serviceProcess) {
+			serviceProcess.kill();
+		}
 	});
 };
